Add Security Policy (#4)
# Pull Request

## Description

This pull request introduces a new security policy to the project
documentation. The most important change is the addition of the
`SECURITY.md` file, which outlines the procedures for reporting
vulnerabilities and details on security updates.

Documentation updates:

*
[`docs/SECURITY.md`](diffhunk://#diff-2605206b7e64be69186fabc8e8fed35676e0015832ff630141aa4aecd73354e0R1-R25):
Added a new security policy document that includes instructions for
reporting vulnerabilities, supported versions for security patches, and
methods for notifying users about security updates.