Sync Repository Labels (#9)
# Pull Request

## Description

This pull request introduces a new labeling system for GitHub issues and
pull requests, and sets up a workflow to automatically sync these
labels. The key changes include the addition of a configuration file for
labels and a GitHub Actions workflow to manage label synchronization.

Labeling system:

*
[`.github/other-configurations/labels.yml`](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dR1-R68):
Added a comprehensive set of default labels for categorizing issues and
pull requests, including labels for bug reports, documentation,
dependencies, enhancements, and more.

Workflow setup:

*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6R1-R27):
Created a GitHub Actions workflow to automatically sync labels based on
the configuration file, ensuring that labels are consistently updated
whenever changes are pushed to the main branch.