Update Repository Labels (#14)
## Description

This pull request includes updates to the labeling configuration and the
list of repositories for label updates. The most important changes
include adding a new label for Markdown files, modifying the paths for
GitHub Actions, and updating the repository list for label operations.

Labeling configuration updates:

*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9R6-R15):
Added a new `markdown` label and updated the file paths to include
additional Markdown files and templates.
*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9L27-R37):
Modified the `github_actions` label to include paths for GitHub actions
in addition to workflows.
*
[`.github/other-configurations/labels.yml`](diffhunk://#diff-c1b32711b0f3da3f5c7007cfaa4d9e94ae5430fea1f4e1256f6210bc6988553dR45-R47):
Added a new label for Markdown documentation with a specific color and
description.

Repository list update:

*
[`scripts/3_update_labels/repos.txt`](diffhunk://#diff-b9326a3a3be08b0929f69b21821f25782056916c24c4f29f1f5d149721b8be0dR1-R18):
Added a list of repositories to be operated upon for label updates.