Schedule Code Checks to Run Daily (#46)
# Pull Request

## Description

This pull request includes a modification to the
`.github/workflows/code-checks.yml` file. The change updates the
workflow name and adjusts the cron schedule for the code checks.

Changes to workflow configuration:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L1-R9):
Updated the workflow name from "Run Code Checks" to "Code Checks" and
changed the cron schedule from "30 1 * * 0" to "0 0 * * *".