Regularly Clear Caches (#51)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow to clean
caches on a scheduled basis. The most important change is the addition
of the `clean-caches.yml` file, which sets up the workflow to run
monthly and allows manual triggering.

New GitHub Actions workflow:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dR1-R24):
Added a new workflow named "Clean Caches" that runs on a monthly
schedule and can be manually triggered. It includes steps to checkout
the repository and delete all caches using the `gh cache delete`
command.