Update Dependency Review (#58)
# Pull Request

## Description

This pull request introduces a new workflow to the GitHub Actions
configuration file `.github/workflows/pull-request-tasks.yml`. The
primary change is the addition of a dependency review job to ensure that
dependencies are reviewed and any issues are highlighted in the pull
request.

Key changes include:

* Added a new job named `dependency-review` to the GitHub Actions
workflow. This job runs on `ubuntu-latest` and has permissions to write
to pull requests.
* The `dependency-review` job includes steps to checkout the repository
and run the dependency review action, configured to comment on the pull
request if there is a failure.