Add 21st Campaign (#77)
# Pull Request

## Description

This pull request includes an update to the `repos.txt` file in the
`scripts/21_update_editorconfig_checker` directory. The change adds
several repositories to the list, all of which are owned by the user
"JackPlowman".

Repository additions:

* Added `JackPlowman/DependabotTrigger` to the list.
* Added `JackPlowman/test-project-status-checker` to the list.
* Added `JackPlowman/windows-development-environment` to the list.
* Added `JackPlowman/repository-template` to the list.
* Added `JackPlowman/RepoSync` to the list.

(Note: Additional repositories were also added, but only the top five
are listed here.)