Add 23rd Campaign (#102)
# Pull Request

## Description

This pull request includes updates to the Lefthook configuration and the
addition of repositories to a Dependabot configuration file. The most
important changes are outlined below:

### Configuration Updates:
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the minimum required version of Lefthook from `1.11.4` to
`1.11.12` to ensure compatibility with newer features or bug fixes.

### Dependabot Configuration:
*
[`scripts/23_dependabot_groups/repos.txt`](diffhunk://#diff-815b205aa3f7f13c1aa20de4628bf610670ae03f255a151643e9a2300f979b0dR1-R29):
Added a list of 29 repositories from the `JackPlowman` namespace to the
Dependabot group configuration, enabling automated dependency updates
for these repositories.