Add 25th Campaign (#107)
# Pull Request

## Description

This pull request introduces a new Zizmor configuration, including
documentation and a list of repositories to be managed. The most
important changes involve the addition of a README file for the
configuration and a comprehensive list of repositories.

### Documentation Updates:
*
[`scripts/25_zizmor_config/README.md`](diffhunk://#diff-0808af7237ddeeac679e572b96a26aef1ac85ba6ea80ea9da54a0fe3d3414236R1-R5):
Added a README file to describe the Zizmor configuration and provide
context for the pull request.

### Repository Management:
*
[`scripts/25_zizmor_config/repos.txt`](diffhunk://#diff-24c31b8f6588853d2772387fa9b286a17b2ed786a750c70935f5982e9a29a0fcR1-R30):
Added a list of repositories to be included in the Zizmor configuration,
covering a variety of projects under the `JackPlowman` namespace.