Add 26th Campaign (#110)
# Pull Request

## Description

This pull request includes updates to the documentation and
configuration for the Zizmor SARIF command. The most important changes
involve adding a description to the `README.md` file and expanding the
list of repositories in `repos.txt`.

### Documentation Updates:
*
[`scripts/26_zizmor_sarif/README.md`](diffhunk://#diff-83fdee64bf1bc617ed4d06349e34449cb68efa45abbb07e94aa44451acd6a973R1-R5):
Added a description section for the Zizmor SARIF command to improve
clarity and provide context.

### Configuration Updates:
*
[`scripts/26_zizmor_sarif/repos.txt`](diffhunk://#diff-37f7d12d811c4e82a9e614c24168d0e2daf7848105712355597836093fefe093R1-R30):
Expanded the list of repositories to include 30 additional entries,
ensuring broader applicability of the Zizmor SARIF command.