Add 28th Campaign (#125)
# Pull Request

## Description

This pull request introduces documentation and configuration updates to
support migrating to a common sync labels system. The key changes
include the addition of a README file describing the migration process
and the inclusion of a list of repositories to be synchronized.

### Documentation updates:
* Added a new `README.md` file in the `scripts/28_common_sync_labels`
directory to provide an overview of the migration process for syncing
labels to a common system.

### Configuration updates:
* Added a `repos.txt` file in the `scripts/28_common_sync_labels`
directory, listing repositories to be included in the common sync labels
process.