Add 29th Campaign (#127)
# Pull Request

## Description

This pull request introduces documentation and configuration updates for
the "Common Clear Caches" script. The most significant changes include
adding a description to the README file and listing repositories in a
new configuration file.

### Documentation Updates:
* Added a description of the "Migrate Clear Caches to Common Clear
Caches" process in the `README.md` file for the `29_common_clear_caches`
script.

### Configuration Updates:
* Created a new `repos.txt` file under the `29_common_clear_caches`
directory, listing repositories to be included in the "Common Clear
Caches" process.