Add 30th Campaign (#129)
# Pull Request

## Description

This pull request updates the documentation and configuration for
Dependabot in reusable workflows. The key changes include adding a new
README section and updating the list of repositories managed by
Dependabot.

### Documentation Update:
* Added a new section titled "Update Dependabot for Reusable Workflows"
in the `README.md` file to describe the purpose and process of the pull
request.

### Configuration Update:
* Updated the `repos.txt` file to include a comprehensive list of
repositories under the `JackPlowman` namespace that Dependabot will
manage.