Add 31st Campaign (#135)
# Pull Request

## Description

This pull request introduces a new script for migrating repositories to
a reusable CodeQL workflow. The changes include documentation for the
migration process and a list of repositories targeted for migration.

### Documentation Updates:
* Added a new `README.md` file in the
`scripts/31_migrate_to_reusable_codeql_workflow` directory to describe
the purpose and details of the migration process.

### Repository List:
* Created a `repos.txt` file in the
`scripts/31_migrate_to_reusable_codeql_workflow` directory, listing 28
repositories to be migrated to the reusable CodeQL workflow.