Add 32th Campaign (#144)
# Pull Request

## Description

This pull request introduces documentation and configuration for the
"Remove CodeLimit Job." It includes a new README file describing the job
and a list of repositories where the job is applicable.

### Documentation Updates:
* Added `README.md` in `scripts/32_Remove_codelimit` to describe the
"Remove CodeLimit Job" and provide context for the pull request.

### Repository Configuration:
* Added `repos.txt` in `scripts/32_Remove_codelimit`, listing 32
repositories where the "Remove CodeLimit Job" will be applied.