Add 33rd Campaign (#145)
# Pull Request

## Description

This pull request introduces documentation and configuration updates for
the "Force Colour in Code Checks Workflow." The changes include adding a
description to the workflow's README and listing repositories in a
configuration file.

Documentation updates:

*
[`scripts/33_Force_Colour/README.md`](diffhunk://#diff-561ff0e746829a22ae30aedd15cfebf227b7c154b933ec9d8379caddd623214fR1-R5):
Added a title and description section for the "Force Colour in Code
Checks Workflow."

Configuration updates:

*
[`scripts/33_Force_Colour/repos.txt`](diffhunk://#diff-9d8e66af5263a806ed0968d7c4d24c41c1411a9a0957d1d43336da1b19272ac4R1-R32):
Added a list of repositories to be included in the workflow
configuration.