Add 34th Campaign (#149)
# Pull Request

## Description

This pull request introduces documentation and configuration updates for
setting workflow job names in repositories. The key changes include the
addition of a README file describing the feature and the inclusion of a
list of repositories where this configuration will be applied.

### Documentation updates:
* Added a new README file (`scripts/34_set_job_name/README.md`) with a
description of the workflow job name feature.

### Configuration updates:
* Created a `repos.txt` file (`scripts/34_set_job_name/repos.txt`)
listing repositories where the workflow job name feature will be
implemented.