Add 35th Campaign (#167)
# Pull Request

## Description

This pull request updates the Lefthook minimum version documentation and
adds a list of repositories for validation. The changes primarily focus
on documentation updates and repository configuration.

### Documentation Updates:
*
[`scripts/35_lefthook_minimum/README.md`](diffhunk://#diff-a049ff9e606136f2af4950343f186f365577c30a0f2067f8eb46c89e208c1a29R1-R5):
Added a description for updating the Lefthook minimum version.

### Repository Configuration:
*
[`scripts/35_lefthook_minimum/repos.txt`](diffhunk://#diff-c47c8f09d55228fbbb3308725ac62fbf413d2ea07d9d368d3c889601f4f06e79R1-R32):
Added a comprehensive list of repositories to be included in the
Lefthook validation process.